Extract some types to be able to reuse them in copilot chat
