+++ src/vs/workbench/api/common/extHostTypes.ts
@@ -6,12 +6,10 @@
 /* eslint-disable local/code-no-native-private */
 
 import type * as vscode from 'vscode';
-import { asArray, coalesceInPlace, equals } from '../../../base/common/arrays.js';
+import { asArray } from '../../../base/common/arrays.js';
 import { VSBuffer } from '../../../base/common/buffer.js';
 import { illegalArgument, SerializedError } from '../../../base/common/errors.js';
 import { IRelativePattern } from '../../../base/common/glob.js';
-import { MarkdownString as BaseMarkdownString, MarkdownStringTrustedOptions } from '../../../base/common/htmlContent.js';
-import { ResourceMap } from '../../../base/common/map.js';
 import { MarshalledId } from '../../../base/common/marshallingIds.js';
 import { Mimes, normalizeMimeType } from '../../../base/common/mime.js';
 import { nextCharLength } from '../../../base/common/strings.js';
@@ -22,36 +20,41 @@ import { TextEditorSelectionSource } from '../../../platform/editor/common/edito
 import { ExtensionIdentifier, IExtensionDescription } from '../../../platform/extensions/common/extensions.js';
 import { FileSystemProviderErrorCode, markAsFileSystemProviderError } from '../../../platform/files/common/files.js';
 import { RemoteAuthorityResolverErrorCode } from '../../../platform/remote/common/remoteAuthorityResolver.js';
-import { CellEditType, ICellMetadataEdit, IDocumentMetadataEdit, isTextStreamMime } from '../../contrib/notebook/common/notebookCommon.js';
+import { isTextStreamMime } from '../../contrib/notebook/common/notebookCommon.js';
 import { IRelativePatternDto } from './extHost.protocol.js';
+import { Position } from './extHostTypes/position.js';
+import { es5ClassCompat } from './extHostTypes/es5ClassCompat.js';
+import { Range } from './extHostTypes/range.js';
+import { CodeActionKind } from './extHostTypes/codeActionKind.js';
+import { Location } from './extHostTypes/location.js';
+import { Diagnostic } from './extHostTypes/diagnostic.js';
+import { TextEdit } from './extHostTypes/textEdit.js';
+import { WorkspaceEdit } from './extHostTypes/workspaceEdit.js';
+import { SnippetString } from './extHostTypes/snippetString.js';
+import { MarkdownString } from './extHostTypes/markdownString.js';
 
-/**
- * @deprecated
- *
- * This utility ensures that old JS code that uses functions for classes still works. Existing usages cannot be removed
- * but new ones must not be added
- * */
-function es5ClassCompat(target: Function): any {
-	const interceptFunctions = {
-		apply: function (...args: any[]): any {
-			if (args.length === 0) {
-				return Reflect.construct(target, []);
-			} else {
-				const argsList = args.length === 1 ? [] : args[1];
-				return Reflect.construct(target, argsList, args[0].constructor);
-			}
-		},
-		call: function (...args: any[]): any {
-			if (args.length === 0) {
-				return Reflect.construct(target, []);
-			} else {
-				const [thisArg, ...restArgs] = args;
-				return Reflect.construct(target, restArgs, thisArg.constructor);
-			}
-		}
-	};
-	return Object.assign(target, interceptFunctions);
-}
+export { Position } from './extHostTypes/position.js';
+export { Range } from './extHostTypes/range.js';
+export { Selection } from './extHostTypes/selection.js';
+export { CodeActionKind } from './extHostTypes/codeActionKind.js';
+export { Location } from './extHostTypes/location.js';
+export {
+	Diagnostic, DiagnosticRelatedInformation,
+	DiagnosticSeverity, DiagnosticTag
+} from './extHostTypes/diagnostic.js';
+export { EndOfLine, TextEdit } from './extHostTypes/textEdit.js';
+export {
+	IFileOperationOptions, FileEditType, IFileOperation,
+	IFileTextEdit, IFileSnippetTextEdit, IFileCellEdit,
+	ICellEdit, WorkspaceEditEntry, WorkspaceEdit
+} from './extHostTypes/workspaceEdit.js';
+export { SnippetString } from './extHostTypes/snippetString.js';
+export { SnippetTextEdit } from './extHostTypes/snippetTextEdit.js';
+export {
+	NotebookCellKind, NotebookRange, NotebookCellData,
+	NotebookData, NotebookEdit
+} from './extHostTypes/notebooks.js';
+export { MarkdownString } from './extHostTypes/markdownString.js';
 
 export enum TerminalOutputAnchor {
 	Top = 0,
@@ -93,430 +96,6 @@ export class Disposable {
 			this.#callOnDispose = undefined;
 		}
 	}
-}
-
-@es5ClassCompat
-export class Position {
-
-	static Min(...positions: Position[]): Position {
-		if (positions.length === 0) {
-			throw new TypeError();
-		}
-		let result = positions[0];
-		for (let i = 1; i < positions.length; i++) {
-			const p = positions[i];
-			if (p.isBefore(result)) {
-				result = p;
-			}
-		}
-		return result;
-	}
-
-	static Max(...positions: Position[]): Position {
-		if (positions.length === 0) {
-			throw new TypeError();
-		}
-		let result = positions[0];
-		for (let i = 1; i < positions.length; i++) {
-			const p = positions[i];
-			if (p.isAfter(result)) {
-				result = p;
-			}
-		}
-		return result;
-	}
-
-	static isPosition(other: any): other is Position {
-		if (!other) {
-			return false;
-		}
-		if (other instanceof Position) {
-			return true;
-		}
-		const { line, character } = <Position>other;
-		if (typeof line === 'number' && typeof character === 'number') {
-			return true;
-		}
-		return false;
-	}
-
-	static of(obj: vscode.Position): Position {
-		if (obj instanceof Position) {
-			return obj;
-		} else if (this.isPosition(obj)) {
-			return new Position(obj.line, obj.character);
-		}
-		throw new Error('Invalid argument, is NOT a position-like object');
-	}
-
-	private _line: number;
-	private _character: number;
-
-	get line(): number {
-		return this._line;
-	}
-
-	get character(): number {
-		return this._character;
-	}
-
-	constructor(line: number, character: number) {
-		if (line < 0) {
-			throw illegalArgument('line must be non-negative');
-		}
-		if (character < 0) {
-			throw illegalArgument('character must be non-negative');
-		}
-		this._line = line;
-		this._character = character;
-	}
-
-	isBefore(other: Position): boolean {
-		if (this._line < other._line) {
-			return true;
-		}
-		if (other._line < this._line) {
-			return false;
-		}
-		return this._character < other._character;
-	}
-
-	isBeforeOrEqual(other: Position): boolean {
-		if (this._line < other._line) {
-			return true;
-		}
-		if (other._line < this._line) {
-			return false;
-		}
-		return this._character <= other._character;
-	}
-
-	isAfter(other: Position): boolean {
-		return !this.isBeforeOrEqual(other);
-	}
-
-	isAfterOrEqual(other: Position): boolean {
-		return !this.isBefore(other);
-	}
-
-	isEqual(other: Position): boolean {
-		return this._line === other._line && this._character === other._character;
-	}
-
-	compareTo(other: Position): number {
-		if (this._line < other._line) {
-			return -1;
-		} else if (this._line > other.line) {
-			return 1;
-		} else {
-			// equal line
-			if (this._character < other._character) {
-				return -1;
-			} else if (this._character > other._character) {
-				return 1;
-			} else {
-				// equal line and character
-				return 0;
-			}
-		}
-	}
-
-	translate(change: { lineDelta?: number; characterDelta?: number }): Position;
-	translate(lineDelta?: number, characterDelta?: number): Position;
-	translate(lineDeltaOrChange: number | undefined | { lineDelta?: number; characterDelta?: number }, characterDelta: number = 0): Position {
-
-		if (lineDeltaOrChange === null || characterDelta === null) {
-			throw illegalArgument();
-		}
-
-		let lineDelta: number;
-		if (typeof lineDeltaOrChange === 'undefined') {
-			lineDelta = 0;
-		} else if (typeof lineDeltaOrChange === 'number') {
-			lineDelta = lineDeltaOrChange;
-		} else {
-			lineDelta = typeof lineDeltaOrChange.lineDelta === 'number' ? lineDeltaOrChange.lineDelta : 0;
-			characterDelta = typeof lineDeltaOrChange.characterDelta === 'number' ? lineDeltaOrChange.characterDelta : 0;
-		}
-
-		if (lineDelta === 0 && characterDelta === 0) {
-			return this;
-		}
-		return new Position(this.line + lineDelta, this.character + characterDelta);
-	}
-
-	with(change: { line?: number; character?: number }): Position;
-	with(line?: number, character?: number): Position;
-	with(lineOrChange: number | undefined | { line?: number; character?: number }, character: number = this.character): Position {
-
-		if (lineOrChange === null || character === null) {
-			throw illegalArgument();
-		}
-
-		let line: number;
-		if (typeof lineOrChange === 'undefined') {
-			line = this.line;
-
-		} else if (typeof lineOrChange === 'number') {
-			line = lineOrChange;
-
-		} else {
-			line = typeof lineOrChange.line === 'number' ? lineOrChange.line : this.line;
-			character = typeof lineOrChange.character === 'number' ? lineOrChange.character : this.character;
-		}
-
-		if (line === this.line && character === this.character) {
-			return this;
-		}
-		return new Position(line, character);
-	}
-
-	toJSON(): any {
-		return { line: this.line, character: this.character };
-	}
-
-	[Symbol.for('debug.description')]() {
-		return `(${this.line}:${this.character})`;
-	}
-}
-
-@es5ClassCompat
-export class Range {
-
-	static isRange(thing: any): thing is vscode.Range {
-		if (thing instanceof Range) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return Position.isPosition((<Range>thing).start)
-			&& Position.isPosition((<Range>thing.end));
-	}
-
-	static of(obj: vscode.Range): Range {
-		if (obj instanceof Range) {
-			return obj;
-		}
-		if (this.isRange(obj)) {
-			return new Range(obj.start, obj.end);
-		}
-		throw new Error('Invalid argument, is NOT a range-like object');
-	}
-
-	protected _start: Position;
-	protected _end: Position;
-
-	get start(): Position {
-		return this._start;
-	}
-
-	get end(): Position {
-		return this._end;
-	}
-
-	constructor(start: vscode.Position, end: vscode.Position);
-	constructor(start: Position, end: Position);
-	constructor(startLine: number, startColumn: number, endLine: number, endColumn: number);
-	constructor(startLineOrStart: number | Position | vscode.Position, startColumnOrEnd: number | Position | vscode.Position, endLine?: number, endColumn?: number) {
-		let start: Position | undefined;
-		let end: Position | undefined;
-
-		if (typeof startLineOrStart === 'number' && typeof startColumnOrEnd === 'number' && typeof endLine === 'number' && typeof endColumn === 'number') {
-			start = new Position(startLineOrStart, startColumnOrEnd);
-			end = new Position(endLine, endColumn);
-		} else if (Position.isPosition(startLineOrStart) && Position.isPosition(startColumnOrEnd)) {
-			start = Position.of(startLineOrStart);
-			end = Position.of(startColumnOrEnd);
-		}
-
-		if (!start || !end) {
-			throw new Error('Invalid arguments');
-		}
-
-		if (start.isBefore(end)) {
-			this._start = start;
-			this._end = end;
-		} else {
-			this._start = end;
-			this._end = start;
-		}
-	}
-
-	contains(positionOrRange: Position | Range): boolean {
-		if (Range.isRange(positionOrRange)) {
-			return this.contains(positionOrRange.start)
-				&& this.contains(positionOrRange.end);
-
-		} else if (Position.isPosition(positionOrRange)) {
-			if (Position.of(positionOrRange).isBefore(this._start)) {
-				return false;
-			}
-			if (this._end.isBefore(positionOrRange)) {
-				return false;
-			}
-			return true;
-		}
-		return false;
-	}
-
-	isEqual(other: Range): boolean {
-		return this._start.isEqual(other._start) && this._end.isEqual(other._end);
-	}
-
-	intersection(other: Range): Range | undefined {
-		const start = Position.Max(other.start, this._start);
-		const end = Position.Min(other.end, this._end);
-		if (start.isAfter(end)) {
-			// this happens when there is no overlap:
-			// |-----|
-			//          |----|
-			return undefined;
-		}
-		return new Range(start, end);
-	}
-
-	union(other: Range): Range {
-		if (this.contains(other)) {
-			return this;
-		} else if (other.contains(this)) {
-			return other;
-		}
-		const start = Position.Min(other.start, this._start);
-		const end = Position.Max(other.end, this.end);
-		return new Range(start, end);
-	}
-
-	get isEmpty(): boolean {
-		return this._start.isEqual(this._end);
-	}
-
-	get isSingleLine(): boolean {
-		return this._start.line === this._end.line;
-	}
-
-	with(change: { start?: Position; end?: Position }): Range;
-	with(start?: Position, end?: Position): Range;
-	with(startOrChange: Position | undefined | { start?: Position; end?: Position }, end: Position = this.end): Range {
-
-		if (startOrChange === null || end === null) {
-			throw illegalArgument();
-		}
-
-		let start: Position;
-		if (!startOrChange) {
-			start = this.start;
-
-		} else if (Position.isPosition(startOrChange)) {
-			start = startOrChange;
-
-		} else {
-			start = startOrChange.start || this.start;
-			end = startOrChange.end || this.end;
-		}
-
-		if (start.isEqual(this._start) && end.isEqual(this.end)) {
-			return this;
-		}
-		return new Range(start, end);
-	}
-
-	toJSON(): any {
-		return [this.start, this.end];
-	}
-
-	[Symbol.for('debug.description')]() {
-		return getDebugDescriptionOfRange(this);
-	}
-}
-
-@es5ClassCompat
-export class Selection extends Range {
-
-	static isSelection(thing: any): thing is Selection {
-		if (thing instanceof Selection) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return Range.isRange(thing)
-			&& Position.isPosition((<Selection>thing).anchor)
-			&& Position.isPosition((<Selection>thing).active)
-			&& typeof (<Selection>thing).isReversed === 'boolean';
-	}
-
-	private _anchor: Position;
-
-	public get anchor(): Position {
-		return this._anchor;
-	}
-
-	private _active: Position;
-
-	public get active(): Position {
-		return this._active;
-	}
-
-	constructor(anchor: Position, active: Position);
-	constructor(anchorLine: number, anchorColumn: number, activeLine: number, activeColumn: number);
-	constructor(anchorLineOrAnchor: number | Position, anchorColumnOrActive: number | Position, activeLine?: number, activeColumn?: number) {
-		let anchor: Position | undefined;
-		let active: Position | undefined;
-
-		if (typeof anchorLineOrAnchor === 'number' && typeof anchorColumnOrActive === 'number' && typeof activeLine === 'number' && typeof activeColumn === 'number') {
-			anchor = new Position(anchorLineOrAnchor, anchorColumnOrActive);
-			active = new Position(activeLine, activeColumn);
-		} else if (Position.isPosition(anchorLineOrAnchor) && Position.isPosition(anchorColumnOrActive)) {
-			anchor = Position.of(anchorLineOrAnchor);
-			active = Position.of(anchorColumnOrActive);
-		}
-
-		if (!anchor || !active) {
-			throw new Error('Invalid arguments');
-		}
-
-		super(anchor, active);
-
-		this._anchor = anchor;
-		this._active = active;
-	}
-
-	get isReversed(): boolean {
-		return this._anchor === this._end;
-	}
-
-	override toJSON() {
-		return {
-			start: this.start,
-			end: this.end,
-			active: this.active,
-			anchor: this.anchor
-		};
-	}
-
-
-	[Symbol.for('debug.description')]() {
-		return getDebugDescriptionOfSelection(this);
-	}
-}
-
-export function getDebugDescriptionOfRange(range: vscode.Range): string {
-	return range.isEmpty
-		? `[${range.start.line}:${range.start.character})`
-		: `[${range.start.line}:${range.start.character} -> ${range.end.line}:${range.end.character})`;
-}
-
-export function getDebugDescriptionOfSelection(selection: vscode.Selection): string {
-	let rangeStr = getDebugDescriptionOfRange(selection);
-	if (!selection.isEmpty) {
-		if (selection.active.isEqual(selection.start)) {
-			rangeStr = `|${rangeStr}`;
-		} else {
-			rangeStr = `${rangeStr}|`;
-		}
-	}
-	return rangeStr;
 }
 
 const validateConnectionToken = (connectionToken: string) => {
@@ -599,621 +178,10 @@ export class RemoteAuthorityResolverError extends Error {
 	}
 }
 
-export enum EndOfLine {
-	LF = 1,
-	CRLF = 2
-}
-
 export enum EnvironmentVariableMutatorType {
 	Replace = 1,
 	Append = 2,
 	Prepend = 3
-}
-
-@es5ClassCompat
-export class TextEdit {
-
-	static isTextEdit(thing: any): thing is TextEdit {
-		if (thing instanceof TextEdit) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return Range.isRange((<TextEdit>thing))
-			&& typeof (<TextEdit>thing).newText === 'string';
-	}
-
-	static replace(range: Range, newText: string): TextEdit {
-		return new TextEdit(range, newText);
-	}
-
-	static insert(position: Position, newText: string): TextEdit {
-		return TextEdit.replace(new Range(position, position), newText);
-	}
-
-	static delete(range: Range): TextEdit {
-		return TextEdit.replace(range, '');
-	}
-
-	static setEndOfLine(eol: EndOfLine): TextEdit {
-		const ret = new TextEdit(new Range(new Position(0, 0), new Position(0, 0)), '');
-		ret.newEol = eol;
-		return ret;
-	}
-
-	protected _range: Range;
-	protected _newText: string | null;
-	protected _newEol?: EndOfLine;
-
-	get range(): Range {
-		return this._range;
-	}
-
-	set range(value: Range) {
-		if (value && !Range.isRange(value)) {
-			throw illegalArgument('range');
-		}
-		this._range = value;
-	}
-
-	get newText(): string {
-		return this._newText || '';
-	}
-
-	set newText(value: string) {
-		if (value && typeof value !== 'string') {
-			throw illegalArgument('newText');
-		}
-		this._newText = value;
-	}
-
-	get newEol(): EndOfLine | undefined {
-		return this._newEol;
-	}
-
-	set newEol(value: EndOfLine | undefined) {
-		if (value && typeof value !== 'number') {
-			throw illegalArgument('newEol');
-		}
-		this._newEol = value;
-	}
-
-	constructor(range: Range, newText: string | null) {
-		this._range = range;
-		this._newText = newText;
-	}
-
-	toJSON(): any {
-		return {
-			range: this.range,
-			newText: this.newText,
-			newEol: this._newEol
-		};
-	}
-}
-
-@es5ClassCompat
-export class NotebookEdit implements vscode.NotebookEdit {
-
-	static isNotebookCellEdit(thing: any): thing is NotebookEdit {
-		if (thing instanceof NotebookEdit) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return NotebookRange.isNotebookRange((<NotebookEdit>thing))
-			&& Array.isArray((<NotebookEdit>thing).newCells);
-	}
-
-	static replaceCells(range: NotebookRange, newCells: NotebookCellData[]): NotebookEdit {
-		return new NotebookEdit(range, newCells);
-	}
-
-	static insertCells(index: number, newCells: vscode.NotebookCellData[]): vscode.NotebookEdit {
-		return new NotebookEdit(new NotebookRange(index, index), newCells);
-	}
-
-	static deleteCells(range: NotebookRange): NotebookEdit {
-		return new NotebookEdit(range, []);
-	}
-
-	static updateCellMetadata(index: number, newMetadata: { [key: string]: any }): NotebookEdit {
-		const edit = new NotebookEdit(new NotebookRange(index, index), []);
-		edit.newCellMetadata = newMetadata;
-		return edit;
-	}
-
-	static updateNotebookMetadata(newMetadata: { [key: string]: any }): NotebookEdit {
-		const edit = new NotebookEdit(new NotebookRange(0, 0), []);
-		edit.newNotebookMetadata = newMetadata;
-		return edit;
-	}
-
-	range: NotebookRange;
-	newCells: NotebookCellData[];
-	newCellMetadata?: { [key: string]: any };
-	newNotebookMetadata?: { [key: string]: any };
-
-	constructor(range: NotebookRange, newCells: NotebookCellData[]) {
-		this.range = range;
-		this.newCells = newCells;
-	}
-}
-
-export class SnippetTextEdit implements vscode.SnippetTextEdit {
-
-	static isSnippetTextEdit(thing: any): thing is SnippetTextEdit {
-		if (thing instanceof SnippetTextEdit) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return Range.isRange((<SnippetTextEdit>thing).range)
-			&& SnippetString.isSnippetString((<SnippetTextEdit>thing).snippet);
-	}
-
-	static replace(range: Range, snippet: SnippetString): SnippetTextEdit {
-		return new SnippetTextEdit(range, snippet);
-	}
-
-	static insert(position: Position, snippet: SnippetString): SnippetTextEdit {
-		return SnippetTextEdit.replace(new Range(position, position), snippet);
-	}
-
-	range: Range;
-
-	snippet: SnippetString;
-
-	keepWhitespace?: boolean;
-
-	constructor(range: Range, snippet: SnippetString) {
-		this.range = range;
-		this.snippet = snippet;
-	}
-}
-
-export interface IFileOperationOptions {
-	readonly overwrite?: boolean;
-	readonly ignoreIfExists?: boolean;
-	readonly ignoreIfNotExists?: boolean;
-	readonly recursive?: boolean;
-	readonly contents?: Uint8Array | vscode.DataTransferFile;
-}
-
-export const enum FileEditType {
-	File = 1,
-	Text = 2,
-	Cell = 3,
-	CellReplace = 5,
-	Snippet = 6,
-}
-
-export interface IFileOperation {
-	readonly _type: FileEditType.File;
-	readonly from?: URI;
-	readonly to?: URI;
-	readonly options?: IFileOperationOptions;
-	readonly metadata?: vscode.WorkspaceEditEntryMetadata;
-}
-
-export interface IFileTextEdit {
-	readonly _type: FileEditType.Text;
-	readonly uri: URI;
-	readonly edit: TextEdit;
-	readonly metadata?: vscode.WorkspaceEditEntryMetadata;
-}
-
-export interface IFileSnippetTextEdit {
-	readonly _type: FileEditType.Snippet;
-	readonly uri: URI;
-	readonly range: vscode.Range;
-	readonly edit: vscode.SnippetString;
-	readonly metadata?: vscode.WorkspaceEditEntryMetadata;
-	readonly keepWhitespace?: boolean;
-}
-
-export interface IFileCellEdit {
-	readonly _type: FileEditType.Cell;
-	readonly uri: URI;
-	readonly edit?: ICellMetadataEdit | IDocumentMetadataEdit;
-	readonly metadata?: vscode.WorkspaceEditEntryMetadata;
-}
-
-export interface ICellEdit {
-	readonly _type: FileEditType.CellReplace;
-	readonly metadata?: vscode.WorkspaceEditEntryMetadata;
-	readonly uri: URI;
-	readonly index: number;
-	readonly count: number;
-	readonly cells: vscode.NotebookCellData[];
-}
-
-
-type WorkspaceEditEntry = IFileOperation | IFileTextEdit | IFileSnippetTextEdit | IFileCellEdit | ICellEdit;
-
-@es5ClassCompat
-export class WorkspaceEdit implements vscode.WorkspaceEdit {
-
-	private readonly _edits: WorkspaceEditEntry[] = [];
-
-
-	_allEntries(): ReadonlyArray<WorkspaceEditEntry> {
-		return this._edits;
-	}
-
-	// --- file
-
-	renameFile(from: vscode.Uri, to: vscode.Uri, options?: { readonly overwrite?: boolean; readonly ignoreIfExists?: boolean }, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.File, from, to, options, metadata });
-	}
-
-	createFile(uri: vscode.Uri, options?: { readonly overwrite?: boolean; readonly ignoreIfExists?: boolean; readonly contents?: Uint8Array | vscode.DataTransferFile }, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.File, from: undefined, to: uri, options, metadata });
-	}
-
-	deleteFile(uri: vscode.Uri, options?: { readonly recursive?: boolean; readonly ignoreIfNotExists?: boolean }, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.File, from: uri, to: undefined, options, metadata });
-	}
-
-	// --- notebook
-
-	private replaceNotebookMetadata(uri: URI, value: Record<string, any>, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.Cell, metadata, uri, edit: { editType: CellEditType.DocumentMetadata, metadata: value } });
-	}
-
-	private replaceNotebookCells(uri: URI, startOrRange: vscode.NotebookRange, cellData: vscode.NotebookCellData[], metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		const start = startOrRange.start;
-		const end = startOrRange.end;
-
-		if (start !== end || cellData.length > 0) {
-			this._edits.push({ _type: FileEditType.CellReplace, uri, index: start, count: end - start, cells: cellData, metadata });
-		}
-	}
-
-	private replaceNotebookCellMetadata(uri: URI, index: number, cellMetadata: Record<string, any>, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.Cell, metadata, uri, edit: { editType: CellEditType.Metadata, index, metadata: cellMetadata } });
-	}
-
-	// --- text
-
-	replace(uri: URI, range: Range, newText: string, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this._edits.push({ _type: FileEditType.Text, uri, edit: new TextEdit(range, newText), metadata });
-	}
-
-	insert(resource: URI, position: Position, newText: string, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this.replace(resource, new Range(position, position), newText, metadata);
-	}
-
-	delete(resource: URI, range: Range, metadata?: vscode.WorkspaceEditEntryMetadata): void {
-		this.replace(resource, range, '', metadata);
-	}
-
-	// --- text (Maplike)
-
-	has(uri: URI): boolean {
-		return this._edits.some(edit => edit._type === FileEditType.Text && edit.uri.toString() === uri.toString());
-	}
-
-	set(uri: URI, edits: ReadonlyArray<TextEdit | SnippetTextEdit>): void;
-	set(uri: URI, edits: ReadonlyArray<[TextEdit | SnippetTextEdit, vscode.WorkspaceEditEntryMetadata | undefined]>): void;
-	set(uri: URI, edits: readonly NotebookEdit[]): void;
-	set(uri: URI, edits: ReadonlyArray<[NotebookEdit, vscode.WorkspaceEditEntryMetadata | undefined]>): void;
-
-	set(uri: URI, edits: null | undefined | ReadonlyArray<TextEdit | SnippetTextEdit | NotebookEdit | [NotebookEdit, vscode.WorkspaceEditEntryMetadata | undefined] | [TextEdit | SnippetTextEdit, vscode.WorkspaceEditEntryMetadata | undefined]>): void {
-		if (!edits) {
-			// remove all text, snippet, or notebook edits for `uri`
-			for (let i = 0; i < this._edits.length; i++) {
-				const element = this._edits[i];
-				switch (element._type) {
-					case FileEditType.Text:
-					case FileEditType.Snippet:
-					case FileEditType.Cell:
-					case FileEditType.CellReplace:
-						if (element.uri.toString() === uri.toString()) {
-							this._edits[i] = undefined!; // will be coalesced down below
-						}
-						break;
-				}
-			}
-			coalesceInPlace(this._edits);
-		} else {
-			// append edit to the end
-			for (const editOrTuple of edits) {
-				if (!editOrTuple) {
-					continue;
-				}
-				let edit: TextEdit | SnippetTextEdit | NotebookEdit;
-				let metadata: vscode.WorkspaceEditEntryMetadata | undefined;
-				if (Array.isArray(editOrTuple)) {
-					edit = editOrTuple[0];
-					metadata = editOrTuple[1];
-				} else {
-					edit = editOrTuple;
-				}
-				if (NotebookEdit.isNotebookCellEdit(edit)) {
-					if (edit.newCellMetadata) {
-						this.replaceNotebookCellMetadata(uri, edit.range.start, edit.newCellMetadata, metadata);
-					} else if (edit.newNotebookMetadata) {
-						this.replaceNotebookMetadata(uri, edit.newNotebookMetadata, metadata);
-					} else {
-						this.replaceNotebookCells(uri, edit.range, edit.newCells, metadata);
-					}
-				} else if (SnippetTextEdit.isSnippetTextEdit(edit)) {
-					this._edits.push({ _type: FileEditType.Snippet, uri, range: edit.range, edit: edit.snippet, metadata, keepWhitespace: edit.keepWhitespace });
-
-				} else {
-					this._edits.push({ _type: FileEditType.Text, uri, edit, metadata });
-				}
-			}
-		}
-	}
-
-	get(uri: URI): TextEdit[] {
-		const res: TextEdit[] = [];
-		for (const candidate of this._edits) {
-			if (candidate._type === FileEditType.Text && candidate.uri.toString() === uri.toString()) {
-				res.push(candidate.edit);
-			}
-		}
-		return res;
-	}
-
-	entries(): [URI, TextEdit[]][] {
-		const textEdits = new ResourceMap<[URI, TextEdit[]]>();
-		for (const candidate of this._edits) {
-			if (candidate._type === FileEditType.Text) {
-				let textEdit = textEdits.get(candidate.uri);
-				if (!textEdit) {
-					textEdit = [candidate.uri, []];
-					textEdits.set(candidate.uri, textEdit);
-				}
-				textEdit[1].push(candidate.edit);
-			}
-		}
-		return [...textEdits.values()];
-	}
-
-	get size(): number {
-		return this.entries().length;
-	}
-
-	toJSON(): any {
-		return this.entries();
-	}
-}
-
-@es5ClassCompat
-export class SnippetString {
-
-	static isSnippetString(thing: any): thing is SnippetString {
-		if (thing instanceof SnippetString) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return typeof (<SnippetString>thing).value === 'string';
-	}
-
-	private static _escape(value: string): string {
-		return value.replace(/\$|}|\\/g, '\\$&');
-	}
-
-	private _tabstop: number = 1;
-
-	value: string;
-
-	constructor(value?: string) {
-		this.value = value || '';
-	}
-
-	appendText(string: string): SnippetString {
-		this.value += SnippetString._escape(string);
-		return this;
-	}
-
-	appendTabstop(number: number = this._tabstop++): SnippetString {
-		this.value += '$';
-		this.value += number;
-		return this;
-	}
-
-	appendPlaceholder(value: string | ((snippet: SnippetString) => any), number: number = this._tabstop++): SnippetString {
-
-		if (typeof value === 'function') {
-			const nested = new SnippetString();
-			nested._tabstop = this._tabstop;
-			value(nested);
-			this._tabstop = nested._tabstop;
-			value = nested.value;
-		} else {
-			value = SnippetString._escape(value);
-		}
-
-		this.value += '${';
-		this.value += number;
-		this.value += ':';
-		this.value += value;
-		this.value += '}';
-
-		return this;
-	}
-
-	appendChoice(values: string[], number: number = this._tabstop++): SnippetString {
-		const value = values.map(s => s.replaceAll(/[|\\,]/g, '\\$&')).join(',');
-
-		this.value += '${';
-		this.value += number;
-		this.value += '|';
-		this.value += value;
-		this.value += '|}';
-
-		return this;
-	}
-
-	appendVariable(name: string, defaultValue?: string | ((snippet: SnippetString) => any)): SnippetString {
-
-		if (typeof defaultValue === 'function') {
-			const nested = new SnippetString();
-			nested._tabstop = this._tabstop;
-			defaultValue(nested);
-			this._tabstop = nested._tabstop;
-			defaultValue = nested.value;
-
-		} else if (typeof defaultValue === 'string') {
-			defaultValue = defaultValue.replace(/\$|}/g, '\\$&'); // CodeQL [SM02383] I do not want to escape backslashes here
-		}
-
-		this.value += '${';
-		this.value += name;
-		if (defaultValue) {
-			this.value += ':';
-			this.value += defaultValue;
-		}
-		this.value += '}';
-
-
-		return this;
-	}
-}
-
-export enum DiagnosticTag {
-	Unnecessary = 1,
-	Deprecated = 2
-}
-
-export enum DiagnosticSeverity {
-	Hint = 3,
-	Information = 2,
-	Warning = 1,
-	Error = 0
-}
-
-@es5ClassCompat
-export class Location {
-
-	static isLocation(thing: any): thing is vscode.Location {
-		if (thing instanceof Location) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return Range.isRange((<Location>thing).range)
-			&& URI.isUri((<Location>thing).uri);
-	}
-
-	uri: URI;
-	range!: Range;
-
-	constructor(uri: URI, rangeOrPosition: Range | Position) {
-		this.uri = uri;
-
-		if (!rangeOrPosition) {
-			//that's OK
-		} else if (Range.isRange(rangeOrPosition)) {
-			this.range = Range.of(rangeOrPosition);
-		} else if (Position.isPosition(rangeOrPosition)) {
-			this.range = new Range(rangeOrPosition, rangeOrPosition);
-		} else {
-			throw new Error('Illegal argument');
-		}
-	}
-
-	toJSON(): any {
-		return {
-			uri: this.uri,
-			range: this.range
-		};
-	}
-}
-
-@es5ClassCompat
-export class DiagnosticRelatedInformation {
-
-	static is(thing: any): thing is DiagnosticRelatedInformation {
-		if (!thing) {
-			return false;
-		}
-		return typeof (<DiagnosticRelatedInformation>thing).message === 'string'
-			&& (<DiagnosticRelatedInformation>thing).location
-			&& Range.isRange((<DiagnosticRelatedInformation>thing).location.range)
-			&& URI.isUri((<DiagnosticRelatedInformation>thing).location.uri);
-	}
-
-	location: Location;
-	message: string;
-
-	constructor(location: Location, message: string) {
-		this.location = location;
-		this.message = message;
-	}
-
-	static isEqual(a: DiagnosticRelatedInformation, b: DiagnosticRelatedInformation): boolean {
-		if (a === b) {
-			return true;
-		}
-		if (!a || !b) {
-			return false;
-		}
-		return a.message === b.message
-			&& a.location.range.isEqual(b.location.range)
-			&& a.location.uri.toString() === b.location.uri.toString();
-	}
-}
-
-@es5ClassCompat
-export class Diagnostic {
-
-	range: Range;
-	message: string;
-	severity: DiagnosticSeverity;
-	source?: string;
-	code?: string | number;
-	relatedInformation?: DiagnosticRelatedInformation[];
-	tags?: DiagnosticTag[];
-
-	constructor(range: Range, message: string, severity: DiagnosticSeverity = DiagnosticSeverity.Error) {
-		if (!Range.isRange(range)) {
-			throw new TypeError('range must be set');
-		}
-		if (!message) {
-			throw new TypeError('message must be set');
-		}
-		this.range = range;
-		this.message = message;
-		this.severity = severity;
-	}
-
-	toJSON(): any {
-		return {
-			severity: DiagnosticSeverity[this.severity],
-			message: this.message,
-			range: this.range,
-			source: this.source,
-			code: this.code,
-		};
-	}
-
-	static isEqual(a: Diagnostic | undefined, b: Diagnostic | undefined): boolean {
-		if (a === b) {
-			return true;
-		}
-		if (!a || !b) {
-			return false;
-		}
-		return a.message === b.message
-			&& a.severity === b.severity
-			&& a.code === b.code
-			&& a.severity === b.severity
-			&& a.source === b.source
-			&& a.range.isEqual(b.range)
-			&& equals(a.tags, b.tags)
-			&& equals(a.relatedInformation, b.relatedInformation, DiagnosticRelatedInformation.isEqual);
-	}
 }
 
 @es5ClassCompat
@@ -1443,51 +411,6 @@ export class CodeAction {
 }
 
 @es5ClassCompat
-export class CodeActionKind {
-	private static readonly sep = '.';
-
-	public static Empty: CodeActionKind;
-	public static QuickFix: CodeActionKind;
-	public static Refactor: CodeActionKind;
-	public static RefactorExtract: CodeActionKind;
-	public static RefactorInline: CodeActionKind;
-	public static RefactorMove: CodeActionKind;
-	public static RefactorRewrite: CodeActionKind;
-	public static Source: CodeActionKind;
-	public static SourceOrganizeImports: CodeActionKind;
-	public static SourceFixAll: CodeActionKind;
-	public static Notebook: CodeActionKind;
-
-	constructor(
-		public readonly value: string
-	) { }
-
-	public append(parts: string): CodeActionKind {
-		return new CodeActionKind(this.value ? this.value + CodeActionKind.sep + parts : parts);
-	}
-
-	public intersects(other: CodeActionKind): boolean {
-		return this.contains(other) || other.contains(this);
-	}
-
-	public contains(other: CodeActionKind): boolean {
-		return this.value === other.value || other.value.startsWith(this.value + CodeActionKind.sep);
-	}
-}
-
-CodeActionKind.Empty = new CodeActionKind('');
-CodeActionKind.QuickFix = CodeActionKind.Empty.append('quickfix');
-CodeActionKind.Refactor = CodeActionKind.Empty.append('refactor');
-CodeActionKind.RefactorExtract = CodeActionKind.Refactor.append('extract');
-CodeActionKind.RefactorInline = CodeActionKind.Refactor.append('inline');
-CodeActionKind.RefactorMove = CodeActionKind.Refactor.append('move');
-CodeActionKind.RefactorRewrite = CodeActionKind.Refactor.append('rewrite');
-CodeActionKind.Source = CodeActionKind.Empty.append('source');
-CodeActionKind.SourceOrganizeImports = CodeActionKind.Source.append('organizeImports');
-CodeActionKind.SourceFixAll = CodeActionKind.Source.append('fixAll');
-CodeActionKind.Notebook = CodeActionKind.Empty.append('notebook');
-
-@es5ClassCompat
 export class SelectionRange {
 
 	range: Range;
@@ -1568,77 +491,6 @@ export class CodeLens {
 
 	get isResolved(): boolean {
 		return !!this.command;
-	}
-}
-
-@es5ClassCompat
-export class MarkdownString implements vscode.MarkdownString {
-
-	readonly #delegate: BaseMarkdownString;
-
-	static isMarkdownString(thing: any): thing is vscode.MarkdownString {
-		if (thing instanceof MarkdownString) {
-			return true;
-		}
-		return thing && thing.appendCodeblock && thing.appendMarkdown && thing.appendText && (thing.value !== undefined);
-	}
-
-	constructor(value?: string, supportThemeIcons: boolean = false) {
-		this.#delegate = new BaseMarkdownString(value, { supportThemeIcons });
-	}
-
-	get value(): string {
-		return this.#delegate.value;
-	}
-	set value(value: string) {
-		this.#delegate.value = value;
-	}
-
-	get isTrusted(): boolean | MarkdownStringTrustedOptions | undefined {
-		return this.#delegate.isTrusted;
-	}
-
-	set isTrusted(value: boolean | MarkdownStringTrustedOptions | undefined) {
-		this.#delegate.isTrusted = value;
-	}
-
-	get supportThemeIcons(): boolean | undefined {
-		return this.#delegate.supportThemeIcons;
-	}
-
-	set supportThemeIcons(value: boolean | undefined) {
-		this.#delegate.supportThemeIcons = value;
-	}
-
-	get supportHtml(): boolean | undefined {
-		return this.#delegate.supportHtml;
-	}
-
-	set supportHtml(value: boolean | undefined) {
-		this.#delegate.supportHtml = value;
-	}
-
-	get baseUri(): vscode.Uri | undefined {
-		return this.#delegate.baseUri;
-	}
-
-	set baseUri(value: vscode.Uri | undefined) {
-		this.#delegate.baseUri = value;
-	}
-
-	appendText(value: string): vscode.MarkdownString {
-		this.#delegate.appendText(value);
-		return this;
-	}
-
-	appendMarkdown(value: string): vscode.MarkdownString {
-		this.#delegate.appendMarkdown(value);
-		return this;
-	}
-
-	appendCodeblock(value: string, language?: string): vscode.MarkdownString {
-		this.#delegate.appendCodeblock(language ?? '', value);
-		return this;
 	}
 }
 
@@ -3853,123 +2705,7 @@ export enum ColorThemeKind {
 }
 
 //#endregion Theming
-
 //#region Notebook
-
-export class NotebookRange {
-	static isNotebookRange(thing: any): thing is vscode.NotebookRange {
-		if (thing instanceof NotebookRange) {
-			return true;
-		}
-		if (!thing) {
-			return false;
-		}
-		return typeof (<NotebookRange>thing).start === 'number'
-			&& typeof (<NotebookRange>thing).end === 'number';
-	}
-
-	private _start: number;
-	private _end: number;
-
-	get start() {
-		return this._start;
-	}
-
-	get end() {
-		return this._end;
-	}
-
-	get isEmpty(): boolean {
-		return this._start === this._end;
-	}
-
-	constructor(start: number, end: number) {
-		if (start < 0) {
-			throw illegalArgument('start must be positive');
-		}
-		if (end < 0) {
-			throw illegalArgument('end must be positive');
-		}
-		if (start <= end) {
-			this._start = start;
-			this._end = end;
-		} else {
-			this._start = end;
-			this._end = start;
-		}
-	}
-
-	with(change: { start?: number; end?: number }): NotebookRange {
-		let start = this._start;
-		let end = this._end;
-
-		if (change.start !== undefined) {
-			start = change.start;
-		}
-		if (change.end !== undefined) {
-			end = change.end;
-		}
-		if (start === this._start && end === this._end) {
-			return this;
-		}
-		return new NotebookRange(start, end);
-	}
-}
-
-export class NotebookCellData {
-
-	static validate(data: NotebookCellData): void {
-		if (typeof data.kind !== 'number') {
-			throw new Error('NotebookCellData MUST have \'kind\' property');
-		}
-		if (typeof data.value !== 'string') {
-			throw new Error('NotebookCellData MUST have \'value\' property');
-		}
-		if (typeof data.languageId !== 'string') {
-			throw new Error('NotebookCellData MUST have \'languageId\' property');
-		}
-	}
-
-	static isNotebookCellDataArray(value: unknown): value is vscode.NotebookCellData[] {
-		return Array.isArray(value) && (<unknown[]>value).every(elem => NotebookCellData.isNotebookCellData(elem));
-	}
-
-	static isNotebookCellData(value: unknown): value is vscode.NotebookCellData {
-		// return value instanceof NotebookCellData;
-		return true;
-	}
-
-	kind: NotebookCellKind;
-	value: string;
-	languageId: string;
-	mime?: string;
-	outputs?: vscode.NotebookCellOutput[];
-	metadata?: Record<string, any>;
-	executionSummary?: vscode.NotebookCellExecutionSummary;
-
-	constructor(kind: NotebookCellKind, value: string, languageId: string, mime?: string, outputs?: vscode.NotebookCellOutput[], metadata?: Record<string, any>, executionSummary?: vscode.NotebookCellExecutionSummary) {
-		this.kind = kind;
-		this.value = value;
-		this.languageId = languageId;
-		this.mime = mime;
-		this.outputs = outputs ?? [];
-		this.metadata = metadata;
-		this.executionSummary = executionSummary;
-
-		NotebookCellData.validate(this);
-	}
-}
-
-export class NotebookData {
-
-	cells: NotebookCellData[];
-	metadata?: { [key: string]: any };
-
-	constructor(cells: NotebookCellData[]) {
-		this.cells = cells;
-	}
-}
-
 
 export class NotebookCellOutputItem {
 
@@ -4095,11 +2831,6 @@ export class CellErrorStackFrame {
 		public uri?: vscode.Uri,
 		public position?: Position,
 	) { }
-}
-
-export enum NotebookCellKind {
-	Markup = 1,
-	Code = 2
 }
 
 export enum NotebookCellExecutionState {
