+++ src/vs/workbench/api/common/extHostTypes/diagnostic.ts
@@ -1,0 +1,107 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { equals } from '../../../../base/common/arrays.js';
+import { URI } from '../../../../base/common/uri.js';
+import { es5ClassCompat } from './es5ClassCompat.js';
+import { Location } from './location.js';
+import { Range } from './range.js';
+
+export enum DiagnosticTag {
+	Unnecessary = 1,
+	Deprecated = 2
+}
+
+export enum DiagnosticSeverity {
+	Hint = 3,
+	Information = 2,
+	Warning = 1,
+	Error = 0
+}
+
+@es5ClassCompat
+export class DiagnosticRelatedInformation {
+
+	static is(thing: any): thing is DiagnosticRelatedInformation {
+		if (!thing) {
+			return false;
+		}
+		return typeof (<DiagnosticRelatedInformation>thing).message === 'string'
+			&& (<DiagnosticRelatedInformation>thing).location
+			&& Range.isRange((<DiagnosticRelatedInformation>thing).location.range)
+			&& URI.isUri((<DiagnosticRelatedInformation>thing).location.uri);
+	}
+
+	location: Location;
+	message: string;
+
+	constructor(location: Location, message: string) {
+		this.location = location;
+		this.message = message;
+	}
+
+	static isEqual(a: DiagnosticRelatedInformation, b: DiagnosticRelatedInformation): boolean {
+		if (a === b) {
+			return true;
+		}
+		if (!a || !b) {
+			return false;
+		}
+		return a.message === b.message
+			&& a.location.range.isEqual(b.location.range)
+			&& a.location.uri.toString() === b.location.uri.toString();
+	}
+}
+
+@es5ClassCompat
+export class Diagnostic {
+
+	range: Range;
+	message: string;
+	severity: DiagnosticSeverity;
+	source?: string;
+	code?: string | number;
+	relatedInformation?: DiagnosticRelatedInformation[];
+	tags?: DiagnosticTag[];
+
+	constructor(range: Range, message: string, severity: DiagnosticSeverity = DiagnosticSeverity.Error) {
+		if (!Range.isRange(range)) {
+			throw new TypeError('range must be set');
+		}
+		if (!message) {
+			throw new TypeError('message must be set');
+		}
+		this.range = range;
+		this.message = message;
+		this.severity = severity;
+	}
+
+	toJSON(): any {
+		return {
+			severity: DiagnosticSeverity[this.severity],
+			message: this.message,
+			range: this.range,
+			source: this.source,
+			code: this.code,
+		};
+	}
+
+	static isEqual(a: Diagnostic | undefined, b: Diagnostic | undefined): boolean {
+		if (a === b) {
+			return true;
+		}
+		if (!a || !b) {
+			return false;
+		}
+		return a.message === b.message
+			&& a.severity === b.severity
+			&& a.code === b.code
+			&& a.severity === b.severity
+			&& a.source === b.source
+			&& a.range.isEqual(b.range)
+			&& equals(a.tags, b.tags)
+			&& equals(a.relatedInformation, b.relatedInformation, DiagnosticRelatedInformation.isEqual);
+	}
+}
